microservices/shared/requirements.txt
# Shared dependencies for Aurora microservices

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
alembic==1.12.1

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Configuration and environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and serialization
email-validator==2.1.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Service discovery (optional)
python-consul==1.1.0

# Message queue (optional)
celery==5.3.4
kombu==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3